Make presenters presentable
Sometime when dealing with an object we don't realy know or care which
class or kind that object is. This can be specially true when dealing
with objects inside views or partials. To avoid major problems this
commit added ApplicationPresenter#present method that returns the presenter
itself so clients can present objects without even worry about whether
or not the object receiving #present is a present itself.